Extract BaseClient that will be shared between airbrake-js and node-airbrake
